+++ app/Http/Controllers/Admin/DocumentController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use App\Model\Admin\Document;
 use MercurySeries\Flashy\Flashy;
 use App\Http\Controllers\Controller;
-
+use Illuminate\Support\Facades\Auth;
 class DocumentController extends Controller
 {
 
@@ -21,13 +21,21 @@ class DocumentController extends Controller
     }
 
     public function index() {
-        $document_all = Document::all();
-        return view('admin.document.index',compact('document_all'));
+        if (Auth::guard('admin')->user()->can('posts.viewAny')) 
+        {
+            $document_all = Document::all();
+            return view('admin.document.index',compact('document_all'));
+        }
+         return redirect(route('admin.home'));
     }
     
     public function create()
     {
-        return view('admin.document.add');
+        if (Auth::guard('admin')->user()->can('posts.create')) 
+        {
+            return view('admin.document.add');
+        }
+         return redirect(route('admin.home'));
     }
 
     /**
@@ -38,28 +46,30 @@ class DocumentController extends Controller
      */
     public function store(Request $request)
     {
-        $validator = $this->validate($request,[
-            'libele' => 'required|string',
-            'type' => 'required|string',
-            'image' => 'required|mimes:pdf,PDF',
-            'resume' => 'required|string',
-        ]);
+        if (Auth::guard('admin')->user()->can('posts.create')) 
+        {
+            $validator = $this->validate($request,[
+                'libele' => 'required|string',
+                'type' => 'required|string',
+                'image' => 'required|mimes:pdf,PDF',
+                'resume' => 'required|string',
+            ]);
 
-        $add_doc = new Document();
-        if ($request->hasFile('image')) {
-            $imageName = $request->image->store('public/Document');
+            $add_doc = new Document();
+            if ($request->hasFile('image')) {
+                $imageName = $request->image->store('public/Document');
+            }
+
+            $add_doc->libele = $request->libele;
+            $add_doc->type = $request->type;
+            $add_doc->resume = $request->resume;
+            $add_doc->status = $request->status;
+            $add_doc->image = $imageName;
+            $add_doc->save();
+            Flashy::success('Votre Document a ete ajoute');
+            return redirect()->route('admin.document.index');
         }
-
-        $add_doc->libele = $request->libele;
-        $add_doc->type = $request->type;
-        $add_doc->resume = $request->resume;
-        $add_doc->status = $request->status;
-        $add_doc->image = $imageName;
-        $add_doc->save();
-        Flashy::success('Votre Document a ete ajoute');
-        return redirect()->route('admin.document.index');
-
-
+         return redirect(route('admin.home'));
     }
 
     /**
@@ -81,8 +91,12 @@ class DocumentController extends Controller
      */
     public function edit($id)
     {
-        $edit_doc = Document::find($id);
-        return view('admin.document.edite',compact('edit_doc'));
+        if (Auth::guard('admin')->user()->can('posts.update')) 
+        {
+            $edit_doc = Document::find($id);
+            return view('admin.document.edite',compact('edit_doc'));
+        }
+         return redirect(route('admin.home'));
     }
 
     /**
@@ -94,20 +108,24 @@ class DocumentController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $update_doc = Document::find($id);
-        if ($request->hasFile('image')) {
-            $imageName = $request->image->store('public/Document');
-        }elseif ($request->image == Null){
-            $imageName = $update_doc->image;
+        if (Auth::guard('admin')->user()->can('posts.update')) 
+        {
+            $update_doc = Document::find($id);
+            if ($request->hasFile('image')) {
+                $imageName = $request->image->store('public/Document');
+            }elseif ($request->image == Null){
+                $imageName = $update_doc->image;
+            }
+            $update_doc->libele = $request->libele;
+            $update_doc->type = $request->type;
+            $update_doc->resume = $request->resume;
+            $update_doc->status = $request->status;
+            $update_doc->image = $imageName;
+            $update_doc->save();
+            Flashy::success('Votre Document a ete modifier');
+            return redirect()->route('admin.document.index');
         }
-        $update_doc->libele = $request->libele;
-        $update_doc->type = $request->type;
-        $update_doc->resume = $request->resume;
-        $update_doc->status = $request->status;
-        $update_doc->image = $imageName;
-        $update_doc->save();
-        Flashy::success('Votre Document a ete modifier');
-        return redirect()->route('admin.document.index');
+         return redirect(route('admin.home'));
     }
 
     /**
@@ -118,9 +136,13 @@ class DocumentController extends Controller
      */
     public function destroy($id)
     {
-        Document::find($id)->delete();
-        Flashy::success('Votre document a ete supprimer');
-        return back();
+        if (Auth::guard('admin')->user()->can('posts.delete')) 
+        {
+            Document::find($id)->delete();
+            Flashy::success('Votre document a ete supprimer');
+            return back();
+        }
+         return redirect(route('admin.home'));
     }
