+++ app/Http/Controllers/Admin/PartenerController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use App\Model\Admin\Partenaire;
 use MercurySeries\Flashy\Flashy;
 use App\Http\Controllers\Controller;
-
+use Illuminate\Support\Facades\Auth;
 class PartenerController extends Controller
 {
     /**
@@ -42,6 +42,8 @@ class PartenerController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::guard('admin')->user()->can('admins.create')) 
+        {
         $this->validate($request,[
             'name' => 'required|string',
             'lien' => 'required|string',
@@ -61,6 +63,8 @@ class PartenerController extends Controller
         $add_partener->save();
         Flashy::success('Votre partenaire a ete ajouter');
         return back();
+        }
+            return redirect(route('admin.home'));
     }
 
     /**
@@ -94,6 +98,8 @@ class PartenerController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
         $this->validate($request,[
             'name' => 'required|string',
             'lien' => 'required|string',
@@ -116,7 +122,10 @@ class PartenerController extends Controller
         $update_partener->save();
         Flashy::success('Votre partenaire a ete modifier');
         return back();
+        }
+        return redirect(route('admin.home'));
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -126,8 +135,12 @@ class PartenerController extends Controller
      */
     public function destroy($id)
     {
+        if (Auth::guard('admin')->user()->can('admins.delete')) 
+        {
         Partenaire::find($id)->delete();
         Flashy::success('Votre Partenaire a ete supprimer');
         return back();
+        }
+        return redirect(route('admin.home'));
     }
 }
