+++ app/Http/Controllers/Admin/PermissionController.php
@@ -1,22 +1,12 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
-
+use Illuminate\Support\Facades\Auth;
 use App\Model\Admin\Permission;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 class PermissionController extends Controller
 {
-
-       /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    // public function __construct()
-    // {
-    //     $this->middleware('auth:admin');
-    // }
 
 
     /**
@@ -31,8 +21,13 @@ class PermissionController extends Controller
     
     public function index()
     {
+        if (Auth::guard('admin')->user()->can('admins.index')) 
+        {
         $permission = Permission::paginate(4);
         return view('admin.permission.index',compact('permission'));
+        }
+        return redirect(route('admin.home'));
+
     }
 
     /**
@@ -42,7 +37,12 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        if (Auth::guard('admin')->user()->can('admins.create')) 
+        {
         return view('admin.permission.create');
+        }
+        return redirect(route('admin.home'));
+
     }
 
     /**
@@ -53,6 +53,8 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::guard('admin')->user()->can('admins.create')) 
+        {
         $this->validate($request,[
             'name' => 'required|max:50',
             'for' => 'required'
@@ -62,6 +64,9 @@ class PermissionController extends Controller
         $permission->for = $request->for;
         $permission->save();
         return redirect(route('admin.permission.index'))->with('message','Permission added succefully');
+        }
+        return redirect(route('admin.home'));
+
     }
 
     /**
@@ -83,8 +88,13 @@ class PermissionController extends Controller
      */
     public function edit($id)
     {
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
         $permission = Permission::find($id);
         return view('admin.permission.edit',compact('permission'));
+        }
+        return redirect(route('admin.home'));
+
     }
 
     /**
@@ -96,6 +106,8 @@ class PermissionController extends Controller
      */
     public function update(Request $request,$id)
     {
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
         $this->validate($request,[
             'name' => 'required|max:50',
             'for' => 'required|max:50'
@@ -105,6 +117,9 @@ class PermissionController extends Controller
         $permission->for = $request->for;
         $permission->save();
         return redirect(route('admin.permission.index'))->with('message','Permission updated succefully');
+        }
+        return redirect(route('admin.home'));
+
     }
 
     /**
@@ -115,7 +130,11 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
+        if (Auth::guard('admin')->user()->can('admins.delete')) 
+        {
         Permission::where('id',$id)->delete();
         return redirect()->back()->with('message','Permission deleted succefully');
+        }
+        return redirect(route('admin.home'));
     }
 }
