+++ app/Http/Controllers/Admin/TeamController.php
@@ -22,9 +22,13 @@ class TeamController extends Controller
     
     public function index()
     {
-        $commission = Commission::all();
-        $teams = Team::all();
-        return view('admin.personnel.index',compact('commission','teams'));
+        if (Auth::guard('admin')->user()->can('admins.index')) 
+        {
+            $commission = Commission::all();
+            $teams = Team::all();
+            return view('admin.personnel.index',compact('commission','teams'));
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -45,27 +49,31 @@ class TeamController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request->All());
-        $validator = $this->validate($request,[
-            'nom' => 'required|string',
-            'email' => 'required|email',
-            'phone' => 'required|numeric',
-            'image' => 'required|image',
-            'poste' => 'required|numeric',
-        ]);
-        if($request->hasFile('image')){
-            $imageName = $request->image->store('public/Personnel');
-        }
+        if (Auth::guard('admin')->user()->can('admins.create')) 
+        {
+            // dd($request->All());
+            $validator = $this->validate($request,[
+                'nom' => 'required|string',
+                'email' => 'required|email',
+                'phone' => 'required|numeric',
+                'image' => 'required|image',
+                'poste' => 'required|numeric',
+            ]);
+            if($request->hasFile('image')){
+                $imageName = $request->image->store('public/Personnel');
+            }
 
-        $add_personnel = new Team;
-        $add_personnel->nom = $request->nom;
-        $add_personnel->email = $request->email;
-        $add_personnel->phone = $request->phone;
-        $add_personnel->poste_id = $request->poste;
-        $add_personnel->image = $imageName;
-        $add_personnel->save();
-        Flashy::success('Votre Personnelle a ete ajouter');
-        return back();
+            $add_personnel = new Team;
+            $add_personnel->nom = $request->nom;
+            $add_personnel->email = $request->email;
+            $add_personnel->phone = $request->phone;
+            $add_personnel->poste_id = $request->poste;
+            $add_personnel->image = $imageName;
+            $add_personnel->save();
+            Flashy::success('Votre Personnelle a ete ajouter');
+            return back();
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -99,20 +107,24 @@ class TeamController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $update_personnel = Team::find($id);
-        $update_personnel->nom = $request->nom;
-        $update_personnel->email = $request->email;
-        $update_personnel->phone = $request->phone;
-        $update_personnel->poste_id = $request->poste;
-        if($request->hasFile('image')){
-            $imageName = $request->image->store('public/Personnel');
-        }else {
-            $imageName = $update_personnel->image;
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
+            $update_personnel = Team::find($id);
+            $update_personnel->nom = $request->nom;
+            $update_personnel->email = $request->email;
+            $update_personnel->phone = $request->phone;
+            $update_personnel->poste_id = $request->poste;
+            if($request->hasFile('image')){
+                $imageName = $request->image->store('public/Personnel');
+            }else {
+                $imageName = $update_personnel->image;
+            }
+            $update_personnel->image = $imageName;
+            $update_personnel->save();
+            Flashy::success('Votre Personnelle a ete modifier');
+            return back();
         }
-        $update_personnel->image = $imageName;
-        $update_personnel->save();
-        Flashy::success('Votre Personnelle a ete modifier');
-        return back();
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -123,8 +135,12 @@ class TeamController extends Controller
      */
     public function destroy($id)
     {
-        Team::find($id)->delete();
-        Flashy::primary('Votre Personnelle a ete Supprimer');
-        return back();
+        if (Auth::guard('admin')->user()->can('admins.delete')) 
+        {
+            Team::find($id)->delete();
+            Flashy::primary('Votre Personnelle a ete Supprimer');
+            return back();
+        }
+        return redirect(route('admin.home'));
     }
 }
