+++ app/Http/Controllers/Admin/TemoignageController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Model\User\Temoignage;
 use Illuminate\Http\Request;
 use MercurySeries\Flashy\Flashy;
+use Illuminate\Support\Facades\Auth;
 class TemoignageController extends Controller
 {
     /**
@@ -15,8 +16,12 @@ class TemoignageController extends Controller
      */
     public function index()
     {
+        if (Auth::guard('admin')->user()->can('admins.index')) 
+        {
         $temoignages = Temoignage::all();
         return view('admin.temoignage.index',compact('temoignages'));
+        }
+         return redirect(route('admin.home'));
     }
 
     /**
@@ -71,11 +76,15 @@ class TemoignageController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
         $update_temoignage = Temoignage::find($id);
         $update_temoignage->status = $request->status;
         $update_temoignage->save();
         Flashy::success('Votre status a ete modifier avec successs');
         return back();
+        }
+         return redirect(route('admin.home'));
     }
 
     /**
@@ -86,8 +95,12 @@ class TemoignageController extends Controller
      */
     public function destroy($id)
     {
+        if (Auth::guard('admin')->user()->can('admins.delete')) 
+        {
         Temoignage::find($id)->delete();
         Flashy::error('Votre status a ete supprimer avec successs');
         return back();
+        }
+         return redirect(route('admin.home'));
     }
 }
