+++ app/Http/Controllers/Admin/UserController.php
@@ -7,7 +7,7 @@ use App\Model\Admin\Role;
 use App\Model\Admin\Admin;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-
+use Illuminate\Support\Facades\Auth;
 class UserController extends Controller
 {
 
@@ -24,8 +24,12 @@ class UserController extends Controller
     
     public function index()
     {
+        if (Auth::guard('admin')->user()->can('admins.index')) 
+        {
         $users = Admin::all();
         return view('admin.user.index',compact('users'));
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -35,8 +39,12 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (Auth::guard('admin')->user()->can('admins.create')) 
+        {
         $roles = Role::all();
         return view('admin.user.create',compact('roles'));
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -47,6 +55,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        if (Auth::guard('admin')->user()->can('admins.create')) 
+        {
         $this->validate($request,[
             'name' => 'required|string',
             'email' => 'required|unique:admins',
@@ -57,6 +67,8 @@ class UserController extends Controller
         $user = Admin::create($request->all());
         $user->roles()->sync($request->role);
         return redirect(route('user.index'));
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -78,9 +90,13 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
         $users = Admin::find($id);
         $roles = Role::all();
         return view('admin.user.edit',compact(['users','roles']));
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -92,6 +108,8 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (Auth::guard('admin')->user()->can('admins.update')) 
+        {
         $this->validate($request,[
             'name' => 'required|string',
             'email' => 'required',
@@ -101,6 +119,8 @@ class UserController extends Controller
         $user = Admin::where('id',$id)->update($request->except('_token','_method','role'));
         Admin::find($id)->roles()->sync($request->role);
         return redirect(route('user.index'))->with('message','Admin updated succuffly');
+        }
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -111,7 +131,11 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (Auth::guard('admin')->user()->can('admins.delete')) 
+        {
         Admin::where('id',$id)->delete();
         return redirect()->back()->with('message','Admin deleted succuffly');
+        }
+        return redirect(route('admin.home'));
     }
 }
